% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carreras_peruanas}
\alias{carreras_peruanas}
\title{Catalog of Peruvian University Degrees}
\format{
A tibble with 93 rows and 2 columns:
\describe{
\item{Carrera}{Standard name of the degree.}
\item{Facultad}{Faculty the degree belongs to.}
}
}
\source{
Compiled by the package author from the degree offer of Peruvian
universities.
}
\usage{
carreras_peruanas
}
\description{
Reference list of undergraduate degrees and their faculties, used by
\code{\link[=normalize_carreras]{normalize_carreras()}} to standardize free-text degree names.
}
\keyword{datasets}
