% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThesiStats-package.R
\docType{package}
\name{ThesiStats-package}
\alias{ThesiStats}
\alias{ThesiStats-package}
\title{ThesiStats: Statistical Tools for Quantitative Theses}
\description{
Helpers for the analyses that quantitative theses in the social and
behavioral sciences repeat: renaming and scoring items, recoding Likert
responses, cleaning sociodemographic variables, descriptive statistics,
normality checks, reliability, correlations and group comparisons with
effect sizes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jventural/ThesiStats}
  \item Report bugs at \url{https://github.com/jventural/ThesiStats/issues}
}

}
\author{
\strong{Maintainer}: José Ventura-León \email{jventuraleon@gmail.com} (\href{https://orcid.org/0000-0003-2996-4244}{ORCID})

}
\keyword{internal}
